+++ CHANGELOG.md
@@ -1,7 +1,15 @@
 # Changelog
 
-All notable changes to `realnum` will be documented in this file
+All notable changes to `realnum` will be documented in this file.
 
-## 0.1.2 - 2019-10-29
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-- beta release
+## [0.2.0] - 2019-11-05
+### Changed
+- Altered format() to use code-distortion/options based option values.
+- Changed locale and noBreakWhitespace to be format-settings.
+- Updated documentation.
+
+## [0.1.2] - 2019-10-29
+### Added
+- Beta release